Added slight timing slop to uart_tx_checker needed for multi_uart
As the clock for multi_uart is external, and provided by the test
infrastructure, it's very accurate to the exact baud timings. This
unfortunately means division rounding doesn't always allow enough slop
for the xcore to output. I've added 0.1% slop to the timings which
should more than account for that.