multi_uart_rx_8chan.S was using 1 level of indirection depth too far
This was causing a runtime exception, as `ldaw` was loading a memory
addr rather than the value in that address. This was wrong, and has
been changed to `ldw`

This bug appeared as a result of moving the CRC lookup tables to
locals in order to make multi_uart able to be re-enterant. (It used to
be a global, accessed with `ldawdp`